[router] Integrate native stack with standard navigation
