+++ packages/expo-router/src/fork/native-stack/createNativeStackNavigator.tsx
@@ -1,5 +1,6 @@
 import { isLiquidGlassAvailable } from 'expo-glass-effect';
 import * as React from 'react';
+import { createStandardNavigator } from 'standard-navigation';
 
 import { useClearGuardedRoutes } from '../../layouts/useClearGuardedRoutes';
 import {
@@ -7,26 +8,13 @@ import {
   type InternalNavigationOptions,
 } from '../../navigationParams';
 import {
-  createNavigatorFactory,
-  type EventArg,
-  type NavigatorTypeBagBase,
-  type ParamListBase,
-  type StackActionHelpers,
-  StackActions,
-  type StackNavigationState,
-  StackRouter,
-  type StackRouterOptions,
-  type TypedNavigator,
-  useNavigationBuilder,
-} from '../../react-navigation/native';
-import {
+  NativeStackView,
+  type NativeStackDescriptorMap,
   type NativeStackNavigationEventMap,
   type NativeStackNavigationOptions,
-  type NativeStackNavigationProp,
-  NativeStackView,
-  type NativeStackNavigatorProps,
-  makePopAction,
 } from '../../react-navigation/native-stack';
+import type { NativeStackNavigationConfig } from '../../react-navigation/native-stack/types';
+import type { NavigatorContentProps } from '../../standard-navigation';
 import { CompositionContext, mergeOptions, useCompositionRegistry } from './composition-options';
 import { DescriptorsContext } from './descriptors-context';
 import { usePreviewTransition } from './usePreviewTransition';
@@ -36,84 +24,49 @@ const GLASS = isLiquidGlassAvailable();
 type NativeStackNavigationOptionsWithInternal = NativeStackNavigationOptions &
   InternalNavigationOptions;
 
-function NativeStackNavigator({
-  id,
-  initialRouteName,
-  children,
-  layout,
-  screenListeners,
-  screenOptions,
-  screenLayout,
-  UNSTABLE_router,
-  ...rest
-}: NativeStackNavigatorProps) {
-  const { state, descriptors, navigation, NavigationContent } = useNavigationBuilder<
-    StackNavigationState<ParamListBase>,
-    StackRouterOptions,
-    StackActionHelpers<ParamListBase>,
-    NativeStackNavigationOptionsWithInternal,
-    NativeStackNavigationEventMap
-  >(StackRouter, {
-    id,
-    initialRouteName,
-    children,
-    layout,
-    screenListeners,
-    screenOptions,
-    screenLayout,
-    UNSTABLE_router,
-  });
+export interface NativeStackNavigatorCreateProps {
+  pop: (count: number, sourceRouteKey: string) => void;
+  removeRoutes: (routeNames: string[]) => void;
+  /** Registers pop-to-top on parent tab press, or returns undefined without a tab parent. */
+  subscribePopToTopOnParentTabPress: () => (() => void) | undefined;
+}
 
-  const removeRoutesFromState = React.useCallback(
-    (routeNames: string[]) =>
-      navigation.dispatch({ type: 'REMOVE_ROUTES', payload: { routeNames } }),
-    [navigation]
-  );
-  useClearGuardedRoutes(removeRoutesFromState);
+export type StandardNativeStackEventMap = {
+  [Event in keyof NativeStackNavigationEventMap]: NativeStackNavigationEventMap[Event] & {
+    canPreventDefault: false;
+  };
+};
 
-  React.useEffect(
-    () =>
-      // @ts-expect-error: there may not be a tab navigator in parent
-      navigation?.addListener?.('tabPress', (e: any) => {
-        const isFocused = navigation.isFocused();
+type ContentArgs = NavigatorContentProps<
+  NativeStackNavigationOptions,
+  StandardNativeStackEventMap,
+  NativeStackNavigationConfig,
+  NativeStackNavigatorCreateProps
+>;
 
-        // Run the operation in the next frame so we're sure all listeners have been run
-        // This is necessary to know if preventDefault() has been called
-        requestAnimationFrame(() => {
-          if (state.index > 0 && isFocused && !(e as EventArg<'tabPress', true>).defaultPrevented) {
-            // When user taps on already focused tab and we're inside the tab,
-            // reset the stack to replicate native behaviour
-            // START FORK
-            // navigation.dispatch({
-            //   ...StackActions.popToTop(),
-            //   target: state.key,
-            // });
-            // The popToTop will be automatically triggered on native side for native tabs
-            if (e.data?.__internalTabsType !== 'native') {
-              navigation.dispatch({
-                ...StackActions.popToTop(),
-                target: state.key,
-              });
-            }
-            // END FORK
-          }
-        });
-      }),
-    [navigation, state.index, state.key]
-  );
+function NativeStackNavigatorContent({
+  state,
+  descriptors,
+  emitter,
+  pop,
+  removeRoutes,
+  subscribePopToTopOnParentTabPress,
+  unstable_nativeProps,
+}: ContentArgs) {
+  const fullDescriptors = descriptors as unknown as NativeStackDescriptorMap;
 
-  // START FORK
-  const { computedState, emit } = usePreviewTransition(state, navigation.emit);
+  const { computedState, emit } = usePreviewTransition(state, emitter.emit);
 
-  const pop = makePopAction(navigation.dispatch, state.key);
+  useClearGuardedRoutes(removeRoutes);
+  React.useEffect(() => subscribePopToTopOnParentTabPress(), [subscribePopToTopOnParentTabPress]);
 
   // Map internal gesture option to React Navigation's gestureEnabled option
   // This allows Expo Router to override gesture behavior without affecting user settings
   const finalDescriptors = React.useMemo(() => {
     let needsNewMap = false;
-    const result: typeof descriptors = {};
-    for (const key of Object.keys(descriptors)) {
-      const descriptor = descriptors[key]!;
+    const result: NativeStackDescriptorMap = {};
+    for (const key of Object.keys(fullDescriptors)) {
+      const descriptor = fullDescriptors[key]!;
       const options = descriptor.options as NativeStackNavigationOptionsWithInternal;
       const internalGestureEnabled = options?.[INTERNAL_EXPO_ROUTER_GESTURE_ENABLED_OPTION_NAME];
       const needsGestureFix = internalGestureEnabled !== undefined;
@@ -136,56 +89,32 @@ function NativeStackNavigator({
         result[key] = descriptor;
       }
     }
-    return needsNewMap ? result : descriptors;
-  }, [descriptors]);
+    return needsNewMap ? result : fullDescriptors;
+  }, [fullDescriptors]);
   const { registry, contextValue } = useCompositionRegistry();
 
   const mergedDescriptors = React.useMemo(
     () => mergeOptions(finalDescriptors, registry, computedState),
     [finalDescriptors, computedState, registry]
   );
-  // END FORK
 
   return (
-    // START FORK
-    <DescriptorsContext value={descriptors}>
-      {/* END FORK */}
-      <NavigationContent>
-        <CompositionContext value={contextValue}>
-          <NativeStackView
-            {...rest}
-            // START FORK
-            state={computedState}
-            descriptors={mergedDescriptors}
-            emit={emit}
-            pop={pop}
-            // state={state}
-            // navigation={navigation}
-            // descriptors={descriptors}
-            // END FORK
-          />
-        </CompositionContext>
-      </NavigationContent>
-      {/* START FORK */}
+    <DescriptorsContext value={fullDescriptors}>
+      <CompositionContext value={contextValue}>
+        <NativeStackView
+          state={computedState}
+          descriptors={mergedDescriptors}
+          emit={emit}
+          pop={pop}
+          unstable_nativeProps={unstable_nativeProps}
+        />
+      </CompositionContext>
     </DescriptorsContext>
-    // END FORK
   );
 }
 
-export function createNativeStackNavigator<
-  const ParamList extends ParamListBase,
-  const NavigatorID extends string | undefined = undefined,
-  const TypeBag extends NavigatorTypeBagBase = {
-    ParamList: ParamList;
-    NavigatorID: NavigatorID;
-    State: StackNavigationState<ParamList>;
-    ScreenOptions: NativeStackNavigationOptions;
-    EventMap: NativeStackNavigationEventMap;
-    NavigationList: {
-      [RouteName in keyof ParamList]: NativeStackNavigationProp<ParamList, RouteName, NavigatorID>;
-    };
-    Navigator: typeof NativeStackNavigator;
-  },
->(): TypedNavigator<TypeBag> {
-  return createNavigatorFactory(NativeStackNavigator)();
-}
+export const createStandardNativeStackNavigator = createStandardNavigator<
+  NativeStackNavigationOptions,
+  StandardNativeStackEventMap,
+  NativeStackNavigatorCreateProps
+>(NativeStackNavigatorContent);
